% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{add_input_to_history}
\alias{add_input_to_history}
\title{add ggbash executed commands in R history}
\usage{
add_input_to_history(input = "point 2 3")
}
\description{
add ggbash executed commands in R history
}

