% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{advice_on_error}
\alias{advice_on_error}
\title{print useful debug advice according to the given error message}
\usage{
advice_on_error(err_message, raw_input = "gg iris | p Sepal.W Sepal.L")
}
\arguments{
\item{err_message}{A character returned by \code{stop}}

\item{raw_input}{A character given to \code{\link{ggbash}} function}
}
\description{
print useful debug advice according to the given error message
}

