% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{build_geom}
\alias{build_geom}
\title{build a geom (a Layer object) from a ggbash character}
\usage{
build_geom(dataset, ggstr = "p x=2 y=3 colour=4 shape=\\"18\\"",
  showWarn = TRUE)
}
\arguments{
\item{dataset}{A dataframe with attr('ggbash_datasetname').}

\item{ggstr}{A ggbash string of geom name, aesthetics, and non-aesthetics}

\item{showWarn}{Show \code{warning} message when ambiguous partial match.
                Default is TRUE.

\code{build_geom} interprets given a ggbash character (vector)
and build a geom string.

\code{build_geom} is a core function in ggbash library.
It performs partial matches against geom name, column names, and aesthetics
and construct a complete geom as a character.}
}
\value{
A list with the following two fields:
\describe{
    \item{geomstr: }{A string representing a geom.}
    \item{geomstr_verbose: }{geomstr with ggplot2:: modifiers.}
    \item{conf: }{the parsed geom/aes/non-aes specifications.}
}
}
\description{
build a geom (a Layer object) from a ggbash character
}
\examples{
build_geom(iris, 'p 1 2 color=5 shape="13"')

}
\seealso{
\code{\link{ggbash}}, \code{\link{copy_to_clipboard}}
}

