% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{copy a given string to clipboard}
\usage{
copy_to_clipboard(string)
}
\arguments{
\item{string}{a character to be copied}
}
\value{
nothing
}
\description{
\code{copy_to_clipboard} invokes OS-specific routine to copy a character to clipboard.
}
\seealso{
\code{\link{drawgg}}
}

