% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{define_ggbash_constants}
\alias{define_ggbash_constants}
\title{define constant values used in ggbash}
\usage{
define_ggbash_constants()
}
\description{
\code{define_ggbash_constants} has no side effect.
It is similar with the 'const' modifier in C or C++.
}
\details{
One thing to note is \code{define_ggbash_constants} set implicitly
the preference order of geom_name in ggplot2.
For example, 'p' ambiguously matches to \code{\link[ggplot2]{geom_point}}
and \code{\link[ggplot2]{geom_pointrange}},
but ggbash automatically uses \code{\link[ggplot2]{geom_point}}
with a warning message about the ambiguity.
This is a design choice based on the observation that
\code{\link[ggplot2]{geom_point}} is often used
more frequently than \code{\link[ggplot2]{geom_pointrange}}.
In order to use \code{\link[ggplot2]{geom_pointrange}},
at least 6 characters ('pointr') is needed.
}
\seealso{
The preference order is used
         when doing partial match in GgplotParser.
}

