% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{draw_ggplot}
\alias{draw_ggplot}
\title{build a ggplot2 plot and draw it}
\usage{
draw_ggplot(dataset, argv = c("p", "x=2", "y=3", "colour=4", "size=5"))
}
\arguments{
\item{dataset}{a dataframe with attr('ggbash_datasetname').}

\item{argv}{a character vector containing ggplot2 geom and aesthetics specifications.
Typically the return value of \code{\link{split_by_space}}.}
}
\value{
a list with the following two fields:
\describe{
    \item{cmd: }{the \code{eval}uated ggplot2 character.}
    \item{conf: }{the parsed aes specifications.}
}
}
\description{
\code{draw_ggplot} interprets given a ggbash character vector
and build a complete ggplot2 object.
}
\details{
\code{draw_ggplot} is a core function in ggbash library.
It performs partial matches against geom name, column names, and aesthetics
and construct a complete ggplot2 object as a character.
This can be used as a function to write one
}
\examples{
out <- draw_ggplot(dataset = iris,
                   argv = split_by_space("line x=Sepal.W y='Sepal.L"))

# copy the built ggplot2 object (Mac OS X)
cat(out$cmd, file=(con <- pipe("pbcopy", "w")))

}

