% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{drawgg}
\alias{drawgg}
\title{build a complete ggplot2 object and print it}
\usage{
drawgg(dataset = NULL, geom_list = list(list(geomstr =
  "geom_point(aes(x=mpg,y=cyl))", geomstr_verbose =
  "ggplot2::geom_point(ggplot2::aes(x=mpg,y=cyl), colour=\\"blue\\", size=6)",
  conf = list(aes = c("x=mpg", "y=cyl"), non_aes = c("colour=\\"blue\\"",
  "size=6"))), list(geomstr = "geom_line(aes(x=mpg,y=cyl))", geomstr_verbose =
  "ggplot2::geom_line(ggplot2::aes(x=mpg,y=cyl), colour=\\"blue\\", size=6)",
  conf = list(aes = c("x=mpg", "y=cyl"), non_aes = c("colour=\\"blue\\"",
  "size=6")))), doEval = TRUE)
}
\arguments{
\item{dataset}{A dataframe with attr('ggbash_datasetname').}

\item{geom_list}{a list of \code{\link{build_geom}} results.}

\item{doEval}{print the built ggplot object. Default is TRUE. Useful for testthat tests.}
}
\value{
A list with the following two fields:
\describe{
    \item{cmd: }{the \code{eval}uated ggplot2 character.}
    \item{cmd_verbose: }{cmd with ggplot2:: modifiers and \code{labs}.}
    \item{conf: }{the parsed geom/aes/non-aes specifications.}
}
}
\description{
\code{drawgg} interprets given a ggbash character vector
and build a complete ggplot2 object.
}
\details{
\code{drawgg} is a core function in ggbash library.
It performs partial matches against geom name, column names, and aesthetics
and construct a complete ggplot2 object as a character.
This can be used as a function to write one
}
\examples{
geom_list <- list(build_geom(iris, 'p 1 2'), build_geom(iris, 'l 1 2'))
out <- drawgg(dataset = iris, geom_list = geom_list)

copy_to_clipboard(out$cmd)

}
\seealso{
\code{\link{ggbash}}, \code{\link{copy_to_clipboard}}, \code{\link{build_geom}}
}

