% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{find_first}
\alias{find_first}
\title{return the first index which contains a given prefix}
\usage{
find_first(prefix = "si", table = c("x", "y", "size", "shape"))
}
\description{
find_first does a prefix partial matching.
}

