% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{find_first_by_prefix}
\alias{find_first_by_prefix}
\title{return the first index which contains a given prefix}
\usage{
find_first_by_prefix(prefix = "si", table = c("x", "y", "size", "shape"),
  show_warn = TRUE)
}
\arguments{
\item{prefix}{A prefix to be searched}

\item{table}{A character vector (typically aesthetic name list)}

\item{show_warn}{Show warning if matched ambiguously. Default is TRUE.}
}
\value{
An integer representing index
}
\description{
find_first_by_prefix does a prefix partial matching.
ggbash tries to interpolate user's input by one of the following mechanism:
prefix match, partial match, or precedence-based guessing.
}

