% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{get_geom_params}
\alias{get_geom_params}
\title{get geom parameters}
\usage{
get_geom_params(suffix = "point")
}
\arguments{
\item{suffix}{geom suffix}
}
\description{
Some geoms like geom_text has special non-aes fields such as check_overlap.
}
\examples{

\dontrun{ get_geom_params("point") }
# returns "na.rm"

\dontrun{ get_geom_params("text") }
# returns "parse" "check_overlap" "na.rm"

}
\seealso{
\code{\link{get_stat_params}}
}

