% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{get_possible_aes}
\alias{get_possible_aes}
\title{retrieve all aesthetic names for a given geom}
\usage{
get_possible_aes(suffix = "point")
}
\arguments{
\item{suffix}{geom suffix}
}
\description{
retrieve all aesthetic names for a given geom
}
\seealso{
used in \code{\link{drawgg}}.
}

