% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{get_required_aes}
\alias{get_required_aes}
\title{retrieve required aesthetic names for a given geom}
\usage{
get_required_aes(suffix = "point")
}
\description{
retrieve required aesthetic names for a given geom
}
\seealso{
used in \code{\link{drawgg}}.
}

