% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{get_stat_params}
\alias{get_stat_params}
\title{return stat params}
\usage{
get_stat_params(suffix = "smooth")
}
\arguments{
\item{suffix}{geom suffix}
}
\description{
Some geoms such as \code{geom_smooth} or \code{geom_histogram}
often set stat parameters (\code{method="lm"} or \code{binwidth}).
The stat parameters is not stored in \code{geom_*()$geom_params},
cannot be obtained by \code{\link{get_geom_params}}
thus retrieve here by another procedure
}
\examples{

\dontrun{ get_stat_params("histogram") }
# returns "binwidth" "bins" "na.rm" "pad"

\dontrun{ get_stat_params("smooth") }
# returns "na.rm" "method" "formula" "se"

\dontrun{ get_stat_params("violin") }
# returns "trim" "scale" "na.rm"

\dontrun{
for (geom in define_ggbash_constants()$geom_namev)
    message(geom, " ", paste0(get_stat_params(geom), collapse=" "))
}


}
\seealso{
\code{\link{get_geom_params}}
}

