% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_after_equal}
\alias{parse_after_equal}
\title{parse symbols after equal sign}
\usage{
parse_after_equal(after = "1 + Sepal.W^2*3", colnamev = c("Sepal.Width",
  "Sepal.Length", "Species"), show_warn = TRUE)
}
\description{
x=factor(Sepal.W + 1) should be interpreted as x = factor(Sepal.Width + 1).
}

