% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_ggbash_aes}
\alias{parse_ggbash_aes}
\title{convert given ggbash strings into ggplot2 aesthetic specifications}
\usage{
parse_ggbash_aes(i, aesv, must_aesv, all_aesv, colnamev)
}
\description{
convert given ggbash strings into ggplot2 aesthetic specifications
}
\seealso{
used in \code{\link{drawgg}}.
}

