% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_ggbash_non_aes}
\alias{parse_ggbash_non_aes}
\title{convert given ggbash strings into ggplot2 non-aesthetic (constant) specifications}
\usage{
parse_ggbash_non_aes(non_aes = "shape=\\"1\\"", all_aesv, showWarn = TRUE)
}
\arguments{
\item{non_aes}{A character of a non-aesthetic key and value pair}

\item{all_aesv}{A vector of possible aesthetics.}

\item{showWarn}{a flag for printing warning when ambiguous match.
Default is TRUE.}
}
\description{
convert given ggbash strings into ggplot2 non-aesthetic (constant) specifications
}
\seealso{
used in \code{\link{drawgg}}.
all_aesv are built by \code{\link{get_possible_aes}}.
\code{\link{parse_ggbash_aes}}
}

