% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_plot_attributes}
\alias{parse_plot_attributes}
\title{parse given plot settings}
\usage{
parse_plot_attributes(argv = c("png", "\\"myname\\"", "900x640"), conf)
}
\arguments{
\item{argv}{A character vector}

\item{conf}{A list of aesthetic assignments}
}
\description{
parse given plot settings
}

