% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_match.R
\name{partial_unique}
\alias{partial_unique}
\title{Construct partially unique string list}
\usage{
partial_unique(originalv = c("mpg", "cyl", "disp", "hp", "drat"), i = 1)
}
\arguments{
\item{originalv}{a character vector to construct partial strings.}

\item{i}{The smallest character of the resulted partial strings.
Sometimes too small i loses readability
such as in \code{\link{show_dataset_column_indices}}}
}
\value{
a list whose element names are partial strings and element values are original strings.
}
\description{
\code{partial_unique} makes given characters as short as posssible while preserving uniqueness.
}
\details{
\code{partial_unique} is a variant of R's builtin \code{\link[base]{unique}} function.
While \code{link[base]{unique}} returns the non-duplicated set of elements,
\code{partial_unique} returns a list whose element names are partially unique shortest strings,
and its element values are original strings.
This function can display how many characters
are needed to uniquely identify each given character element.
}
\examples{
partial_unique(c("Sepal.Width", "Sepal.Length", "Species", "Petal.Width"))
# returns list(Sepal.W = "Sepal.Width", Sepal.L = "Sepal.Length",
#              Sp = "Species", P = "Petal.Width")

}
\seealso{
\code{\link[base]{unique}}, \code{\link{show_dataset_column_indices}}
}

