% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{remove_aes}
\alias{remove_aes}
\title{remove aes() function call}
\usage{
remove_aes(input = "gg(a) + p(aes(x,y),c) + p(aes(l),d)")
}
\arguments{
\item{input}{A character}
}
\description{
ggbash() tries to follow ggplot2 original syntax
as much as possible.
When aes() function calls is specified in a given
ggbash string, those are safely removed by this function.
}

