% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{remove_unnecessary_marks}
\alias{remove_unnecessary_marks}
\title{remove parentheses and marks}
\usage{
remove_unnecessary_marks(input = "gg(m, x=f(cyl), m) + t(l=p0('l:', w))")
}
\arguments{
\item{input}{A character}
}
\description{
ggbash have to handle two types of parentheses and tyo types of commas.
One is unnecessary and the other is necessary for correct parsing.
For example, in \code{gg(iris) + bin2d(x, y, binwidth=c(.1, .1))},
all parentheses and commas before "binwidth" are unnecessary.
So this should be replaced by
\code{gg iris  + bin2d x  y  binwidth=c(.1, .1) },
which can be parsed by ggplot2 compiler.
}
\details{
This is because while ggbash does not rely on commas and parens for
parsing, R relies on them.
}

