% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save a ggplot object into a file}
\usage{
save_ggplot(exe_statl = list(cmd =
  "ggplot(mtcars) + geom_point(aes(cyl,mpg))", conf = list("x=cyl", "y=mpg")),
  argv = c("png", "big"))
}
\arguments{
\item{exe_statl}{A list resulted from \code{\link{drawgg}}}

\item{argv}{A character vector}
}
\description{
save a ggplot object into a file
}

