% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{set_ggbash_dataset}
\alias{set_ggbash_dataset}
\title{build a data frame from a data frame name}
\usage{
set_ggbash_dataset(dataset_name, quietly = FALSE)
}
\arguments{
\item{dataset_name}{a character representing a data frame}

\item{quietly}{Default is FALSE. Useful for testthat tests}
}
\value{
a tbl_df object with attr('ggbash_datasetname')
}
\description{
\code{set_ggbash_dataset} receives a character (a data frame name),
evaluate it as a symbol, and construct a corresponding tbl_df object.
The given character argument is stored in attr('ggbash_datasetname')
for future reference in \code{\link{drawgg}}.
}
\examples{

newdf <- set_ggbash_dataset('iris')
attr(newdf, 'ggbash_datasetname')  # 'iris'

}
\seealso{
\code{\link{drawgg}}
}

