% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{set_ggplot_default_aes}
\alias{set_ggplot_default_aes}
\title{set ggplot() aesthetics arguments}
\usage{
set_ggplot_default_aes(dataset, aesv = c("x=mpg", "y=cyl", "c=am"))
}
\arguments{
\item{dataset}{A data frame}

\item{aesv}{A vector of aesthetics}
}
\description{
set ggplot() aesthetics arguments
}

