% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{set_layer_colnames}
\alias{set_layer_colnames}
\title{set layer-specific column names}
\usage{
set_layer_colnames(input = "gg(a) + p(x=x, color='red', y=y, data=iris)")
}
\arguments{
\item{input}{A ggbash string}
}
\description{
this functio nset ggbashenv$layer_coll (column name list),
which is used for layer-specific column name partial match.
}

