% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{show_fixit_diagnostics}
\alias{show_fixit_diagnostics}
\title{Display useful debugging info for users}
\usage{
show_fixit_diagnostics(err = list(id = "p_theme_elem:prefix_match", type =
  "Prefix match for theme element name failed.", input = "axis.tx:", elem_name =
  "axis.tx", elem_table = c("axis.text", "axis.title")))
}
\arguments{
\item{err}{A list of error information}
}
\description{
Display useful debugging info for users
}

