% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{split_by_pipe}
\alias{split_by_pipe}
\title{split a given character by a pipe ("|" or "+")}
\usage{
split_by_pipe(input = "point x=3 y=4 color=5 | copy")
}
\arguments{
\item{input}{A character}
}
\description{
One thing to note here is that
ggbash does not differentiate between "|" and "+".
This follows the observation that sometimes
ggplot2 users would tend to mistype
not
mtcars %>% ggplot() + geom_point(aes(mpg,cyl))
but
mtcars %>% ggplot() %>% geom_point(aes(mpg,cyl))
}

