% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{split_by_space}
\alias{split_by_space}
\title{split a given string by spaces}
\usage{
split_by_space(input = "    point x=3 y=4 color=5 ")
}
\arguments{
\item{input}{A character. Typically one of the elements returned by \code{\link{split_by_pipe}}.}
}
\value{
A character vector
}
\description{
split a given string by spaces
}

