% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{theme2}
\alias{theme2}
\title{an enhanced version of ggplot2::theme()}
\usage{
theme2(..., as_string = FALSE)
}
\arguments{
\item{...}{theme element specification (see examples below)}

\item{as_string}{return the built theme function call as string.
Default is FALSE.}
}
\description{
theme2() has an enhanced version of ggplot2::theme() in terms of:
1. no element_(text|line|rect|grob|blank) specification
2. partial match for each configuration (e.g. size by sz)
}
\examples{
\dontrun{

# all of the following three generate the same plot

ggplot(mtcars) + geom_point(aes(wt, hp, color=cyl)) +
   theme(text = element_text(size=20, face="bold"),
   axis.line = element_line(size=2),
   legend.key = element_rect(color="black"))

ggplot(mtcars) + geom_point(aes(wt, hp, color=cyl)) +
   theme2(text(size=20, face="bold"), axis.line(size=2),
          legend.key(color="black"))

ggplot(mtcars) + geom_point(aes(wt, hp, color=cyl)) +
   theme2(text(sz=20, f="bold"), axis.line(sz=2),
          legend.key(c="black"))


}

}

