% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{compile_ggbash}
\alias{compile_ggbash}
\title{the core function of ggbash}
\usage{
compile_ggbash(cmd = "gg mtcars + p wt wt")
}
\arguments{
\item{cmd}{A character

compile_ggbash returns a built ggplot object.}
}
\description{
the core function of ggbash
}

