% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{drawgg}
\alias{drawgg}
\title{build a ggplot2 plot and draw it}
\usage{
drawgg(dataset, argv = c("p", "x=2", "y=3", "colour=4", "size=5"),
  showWarning = TRUE)
}
\arguments{
\item{dataset}{A dataframe with attr('ggbash_datasetname').}

\item{argv}{A character vector containing ggplot2 geom and aesthetics specifications.
Typically the return value of \code{\link{split_by_space}}.}

\item{showWarning}{whether to show warning when ambiguously matched. Default is TRUE.}
}
\value{
A list with the following two fields:
\describe{
    \item{cmd: }{the \code{eval}uated ggplot2 character.}
    \item{conf: }{the parsed aes specifications.}
}
}
\description{
\code{drawgg} interprets given a ggbash character vector
and build a complete ggplot2 object.
}
\details{
\code{drawgg} is a core function in ggbash library.
It performs partial matches against geom name, column names, and aesthetics
and construct a complete ggplot2 object as a character.
This can be used as a function to write one
}
\examples{
out <- drawgg(dataset = iris,
                   argv = split_by_space("line x=Sepal.W y='Sepal.L"))

# copy the built ggplot2 object (Mac OS X)
copy_to_clipboard(out$cmd)

}
\seealso{
\code{\link{ggbash}}, \code{\link{copy_to_clipboard}}
}

