% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{exec_ggbash}
\alias{exec_ggbash}
\title{execute raw ggbash commands}
\usage{
exec_ggbash(dataset, raw_input = "point 1 2 | copy", showWarning = TRUE)
}
\arguments{
\item{dataset}{A dataframe}

\item{raw_input}{A ggbash command chain (might contain pipes)}

\item{showWarning}{Whether to show a warning message
when ambiguously matched. Default is TRUE.}
}
\description{
execute raw ggbash commands
}

