% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_match_utils.R
\name{find_first}
\alias{find_first}
\title{return the first index which contains a given prefix}
\usage{
find_first(prefix = "si", table = c("x", "y", "size", "shape"),
  showWarn = TRUE)
}
\arguments{
\item{prefix}{A prefix to be searched}

\item{table}{A character vector (typically aesthetic name list)}

\item{showWarn}{Show warning if matched ambiguously. Default is TRUE.}
}
\description{
find_first does a prefix partial matching.
}

