% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{ggbash}
\alias{ggbash}
\title{Tell ggbash to execute a given ggbash command}
\usage{
ggbash(ggbash_symbols = "", clipboard = NULL, show_warn = TRUE,
  as_string = FALSE)
}
\arguments{
\item{ggbash_symbols}{Non-evaluated R symbols or
a character representing ggbash commands.
If no ggbash_symbols are specified,
enter into an interactive ggbash session.}

\item{clipboard}{Default is NULL.
If non-null, copy the resulted string to clipboard.}

\item{show_warn}{If ambiguously matched, display warning. Default is TRUE.}

\item{as_string}{Return a string instead of a ggplot2 object.
Default is FALSE.}
}
\description{
\code{ggbash()} is used in the following three different ways:
1. ggbash("gg mtcars + point mpg cyl") : with a character argument
2. ggbash(gg(mtcars) + point(mpg,cyl)) : with a pseudo-ggplot2 command
3. ggbash() : with no argument (enter into an interactive ggbash session)
}
\details{
While ggbash can interpret "with- and without- parentheses" and
"with- and without- commas" commands in 1 and 3,
2. can only interpret the with-parentheses and with-commas case
because of R's default token constraints.
}

