% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{ggbash}
\alias{ggbash}
\title{Enter into a ggbash session.}
\usage{
ggbash(dataset = NULL)
}
\arguments{
\item{dataset}{a dataframe to attach (default is NULL).
You can define later by 'use your_dataset' command in.}
}
\value{
nothing
}
\description{
\code{ggbash} executes a new ggbash session within the current R session.
}
\examples{
ggbash()
ggbash(iris)
}

