% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{ggbash}
\alias{ggbash}
\title{Enter into a ggbash session.}
\usage{
ggbash(showWarn = TRUE)
}
\arguments{
\item{showWarn}{Whether to show a warning message
                   when ambiguously matched. Default is TRUE.
\describe{
    \item{Geom name:}{the geom most frequently used (based on my experiences)}
    \item{Column name:}{the column with the smallest column index}
    \item{Aesthetics:}{required (x, y), non-missing (shape, size), default (alpha, stroke) }
}}
}
\value{
nothing
}
\description{
\code{ggbash} executes a new ggbash session for faster ggplot2 plotting.
}
\details{
ggbash provides concise aliases for ggplot2 functions.
By calling ggbash(), your R session goes into a ggbash session,
which only interprets predefined ggbash commands.
Some basic commands like setwd() or pwd() works in ggbash session,
but most of the usual R grammars are disabled.
Instead, a variety of ggbash commands are enabled
for writing ggplot2 script as faster as possible.
}
\examples{
\dontrun{ ggbash()
}

}
\seealso{
For a oneliner, \code{\link{drawgg}} might be more convenient.
}

