% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{ggbash}
\alias{ggbash}
\title{Enter into a ggbash session.}
\usage{
ggbash(dataset = NULL, ambiguous_match = TRUE, showWarning = TRUE)
}
\arguments{
\item{dataset}{A dataframe to attach. Default is NULL.
You can define the dataframe later
by 'use your_dataset' command in your ggbash session.
If a matrix object is given, It's automatically
converted into a tbl_df object.}

\item{ambiguous_match}{A boolean whether to do ambiguous match when a
ggbash command ambiguously matches several commands.
Default is TRUE. The matching rules are as follows:}

\item{showWarning}{Whether to show a warning message
                   when ambiguously matched. Default is TRUE.
\describe{
    \item{Geom name:}{the geom most frequently used (based on my experiences)}
    \item{Column name:}{the column with the smallest column index}
    \item{Aesthetics:}{required (x, y), non-missing (shape, size), default (alpha, stroke) }
}}
}
\value{
nothing
}
\description{
\code{ggbash} executes a new ggbash session for faster ggplot2 plotting.
}
\details{
ggbash provides concise aliases for ggplot2 functions.
By calling ggbash(), your R session goes into a ggbash session,
which only interprets predefined ggbash commands.
Some basic commands like setwd() or pwd() works in ggbash session,
but most of the usual R grammars are disabled.
Instead, a variety of ggbash commands are enabled
for writing ggplot2 script as faster as possible.
}
\examples{
\dontrun{ ggbash()
ggbash(iris)
}

}
\seealso{
For a oneliner, \code{\link{drawgg}} might be more convenient.
}

