% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_ggbash_aes}
\alias{parse_ggbash_aes}
\title{convert given ggbash strings into ggplot2 aesthetic specifications}
\usage{
parse_ggbash_aes(i, aesv, must_aesv, all_aesv, colnamev, show_warn = TRUE)
}
\arguments{
\item{i}{An integer of index}

\item{aesv}{A vector of aesthetics}

\item{must_aesv}{A vector of required aesthetics}

\item{all_aesv}{A vector of possible aesthetics.}

\item{colnamev}{A vector of column names of a dataframe.}

\item{show_warn}{a flag for printing warning when ambiguous match.
                   Default is TRUE.

must_aesv and all_aesv are built by
\code{\link{get_required_aes}} and
\code{\link{get_possible_aes}}, respectively.}
}
\description{
convert given ggbash strings into ggplot2 aesthetic specifications
}

