% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{parse_plot_attributes}
\alias{parse_plot_attributes}
\title{parse given plot settings}
\usage{
parse_plot_attributes(argv = c("png", "\\"myname\\"", "900x640",
  "my_plot_dir/"), conf = list(aes = list("x=cyl", "y=mpg"), non_aes =
  list("color=\\"blue\\"", "shape=\\"18\\""), geom_list = c("point", "smooth")),
  dataset_string = "mtcars-32")
}
\arguments{
\item{argv}{A character vector}

\item{conf}{A list of aesthetic and non-aesthetic assignments}

\item{dataset_string}{A character representing a dataset directory}
}
\description{
parse given plot settings
}

