% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{save_ggplot}
\alias{save_ggplot}
\title{save a ggplot object into a file}
\usage{
save_ggplot(dataset_string = "mtcars-32",
  ggstr = "ggplot(iris) + geom_point(aes(Sepal.Width, Sepal.Length))",
  conf = list(aes = c("x=cyl", "y=mpg"), non_aes = c(), geom_list = "point"),
  argv = c("png", "200*500", "'my-file-name'", "my_plot_dir"))
}
\arguments{
\item{dataset_string}{A character. Used as a directory.}

\item{ggstr}{A list of aesthetics}

\item{conf}{A list of aesthetics, non-aesthetics and geoms}

\item{argv}{A character vector}
}
\description{
save a ggplot object into a file
}

