% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{show_dataset_column_indices}
\alias{show_dataset_column_indices}
\title{show column index list}
\usage{
show_dataset_column_indices(dataset_str = NULL)
}
\arguments{
\item{dataset_str}{A character representing a data frame}
}
\description{
This function lists all dataset column indices.
}
\seealso{
\code{partial_unique}
}

