% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{show_dataset_column_indices}
\alias{show_dataset_column_indices}
\title{show column index list}
\usage{
show_dataset_column_indices(dataset = NULL)
}
\arguments{
\item{dataset}{a data frame}
}
\description{
This function lists all dataset column indices.
}
\examples{

# without column indices: explicit
drawgg(iris, split_by_space("line x=Sepal.W y=Sepal.L colour=Species"))

show_dataset_column_indices(iris)

# with column indices: shorter
drawgg(iris, split_by_space('l x=2 y=1 c=5'))

}
\seealso{
\code{partial_unique}, \code{drawgg}
}

