% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbash.R
\name{split_by_pipe}
\alias{split_by_pipe}
\title{split a given character by a pipe ("|")}
\usage{
split_by_pipe(input = "point x=3 y=4 color=5 | copy")
}
\arguments{
\item{input}{A character}
}
\description{
split a given character by a pipe ("|")
}

